12 AUG FINAL

+++ Classifier.py
@@ -862,7 +862,7 @@ class Classifier:
         self.newData_=self.newData_.reindex(columns=list(self.FE_.X_train_.columns),fill_value=0)
         #self.prediction_=self.bestModels_.copy()
         for key in self.bestModels_:
-            self.predictionReport_.insertPredictionResults([key,self.bestModels_[key].predict(self.newData_),int(np.round(self.bestModels_[key].predict_proba(self.newData_)[0][0],2)*100),int(np.round(self.bestModels_[key].predict_proba(self.newData_)[0][1],2)*100)])              
+            self.predictionReport_.insertPredictionResults([key,int(self.bestModels_[key].predict(self.newData_)),int(np.round(self.bestModels_[key].predict_proba(self.newData_)[0][0],2)*100),int(np.round(self.bestModels_[key].predict_proba(self.newData_)[0][1],2)*100)])              
         print(self.predictionReport_.predictionReport_)
